@@ -1,21 +1,50 @@
-source "https://rubygems.org"
+source 'https://rubygems.org'
 
-ruby '1.9.3'
-gem 'rails', '2.3.17'
-gem 'rake', '0.8.7'
-gem 'sqlite3', '1.3.3'
 
-gem 'aasm', '2.2.0'
-gem 'authlogic', '2.1.6'
-gem 'super_exception_notifier', '2.0.8', :require => 'exception_notifier'
-gem 'json', '1.5.5'
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
+gem 'rails', '4.1.8'
+# Use sqlite3 as the database for Active Record
+gem 'sqlite3'
+# Use SCSS for stylesheets
+gem 'sass-rails', '~> 4.0.3'
+# Use Uglifier as compressor for JavaScript assets
+gem 'uglifier', '>= 1.3.0'
+# Use CoffeeScript for .js.coffee assets and views
+gem 'coffee-rails', '~> 4.0.0'
+# See https://github.com/sstephenson/execjs#readme for more supported runtimes
+# gem 'therubyracer',  platforms: :ruby
 
-gem 'factory_girl', '1.3.3', :require => false, :groups => ['development', 'test']
-gem 'rspec-rails',  '1.3.3', :require => false, :groups => ['development', 'test']
-gem 'debugger', :groups => ['development', 'test']
-gem 'test-unit',  '1.2.3', :groups => ['development', 'test']
-gem 'rdoc'
+# Use jquery as the JavaScript library
+gem 'jquery-rails'
+# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
+gem 'turbolinks'
+# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
+gem 'jbuilder', '~> 2.0'
+# bundle exec rake doc:rails generates the API under doc/api.
+gem 'sdoc', '~> 0.4.0',          group: :doc
 
-group :development do
-  gem 'capistrano', :require => false
+# Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+gem 'spring',        group: :development
+
+# Use ActiveModel has_secure_password
+# gem 'bcrypt', '~> 3.1.7'
+
+# Use unicorn as the app server
+# gem 'unicorn'
+
+# Use Capistrano for deployment
+gem 'capistrano-rails', group: :development
+
+# Use debugger
+# gem 'debugger', group: [:development, :test]
+
+gem 'rspec-rails',  '~> 3.0', :require => false, :groups => ['development', 'test']
+gem 'factory_girl_rails', '~> 4.0', :require => false, :groups => ['development', 'test']
+gem 'database_cleaner', '~> 1.4.0', :require => false, :groups => ['development', 'test']
+
+gem 'authlogic', '~> 3.0'
+gem 'aasm', '~> 4.1.0'
+
+group :production do
+  gem 'exception_notification'
 end
