@@ -1,25 +1,13 @@
 class TicketsController < ApplicationController
-  before_filter :assign_event, :only => [:index]
+  before_filter :assign_event, only: [:index]
 
   # GET /tickets
-  # GET /tickets.xml
   def index
     @tickets = Ticket.ordered
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.xml  { render :xml => @tickets }
-    end
   end
 
   # GET /tickets/1
-  # GET /tickets/1.xml
   def show
     @ticket = Ticket.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.xml  { render :xml => @ticket }
-    end
   end
 end
