@@ -1,6 +1,6 @@
 class UserSessionsController < ApplicationController
-  before_filter :require_no_user, :only => [:new, :create]
-  before_filter :require_user, :only => :destroy
+  before_filter :require_no_user, only: [:new, :create]
+  before_filter :require_user, only: :destroy
 
   def new
     @user_session = UserSession.new
@@ -12,7 +12,7 @@ class UserSessionsController < ApplicationController
       flash[:notice] = "Login successful!"
       redirect_back_or_default root_url
     else
-      render :action => :new
+      render action: :new
     end
   end
 
