@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe TicketKindsController do
 
@@ -7,99 +7,99 @@ describe TicketKindsController do
   end
 
   def create_record
-    @record = Factory(:ticket_kind)
+    @record = create(:ticket_kind)
   end
 
   def create_records
     @records = [
-      Factory(:ticket_kind),
-      Factory(:ticket_kind),
+      create(:ticket_kind),
+      create(:ticket_kind),
     ]
   end
 
   describe "index" do
-    it "should list ticket kinds" do
+    it "expects to list ticket kinds" do
       TicketKind.destroy_all
       create_records
       get :index
-      response.should be_success
-      flash[:error].should be_blank
-      assigns[:ticket_kinds].should == @records
+      expect(response).to be_success
+      expect(flash[:error]).to be_blank
+      expect(assigns[:ticket_kinds]).to == @records
     end
 
-    it "should list empty set" do
+    it "expects to list empty set" do
       get :index
-      response.should be_success
-      flash[:error].should be_blank
-      assigns[:ticket_kinds].should == []
+      expect(response).to be_success
+      expect(flash[:error]).to be_blank
+      expect(assigns[:ticket_kinds]).to == []
     end
   end
 
   describe "new" do
-    it "should display form" do
+    it "expects to display form" do
       get :new
-      response.should be_success
-      flash[:error].should be_blank
+      expect(response).to be_success
+      expect(flash[:error]).to be_blank
     end
   end
 
   describe "create" do
-    it "should succeed when given valid attributes" do
-      attributes = Factory.attributes_for(:ticket_kind)
+    it "expects to succeed when given valid attributes" do
+      attributes = attributes_for(:ticket_kind)
       post :create, :ticket_kind => attributes
-      response.should redirect_to(ticket_kind_path(assigns[:ticket_kind]))
-      assigns[:ticket_kind].should be_valid
+      expect(response).to redirect_to(ticket_kind_path(assigns[:ticket_kind]))
+      expect(assigns[:ticket_kind]).to be_valid
     end
 
-    it "should fail when given invalid attributes" do
-      attributes = Factory.attributes_for(:ticket_kind, :title => nil)
+    it "expects to fail when given invalid attributes" do
+      attributes = attributes_for(:ticket_kind, :title => nil)
       post :create, :ticket_kind => attributes
-      response.should be_success
-      assigns[:ticket_kind].should_not be_valid
+      expect(response).to be_success
+      expect(assigns[:ticket_kind]).to_not be_valid
     end
   end
 
   describe "show" do
-    it "should succeed" do
+    it "expects to succeed" do
       create_record
       get :show, :id => @record.id
-      response.should be_success
-      assigns[:ticket_kind].should == @record
+      expect(response).to be_success
+      expect(assigns[:ticket_kind]).to == @record
     end
   end
 
   describe "edit" do
-    it "should succeed" do
+    it "expects to succeed" do
       create_record
       get :edit, :id => @record.id
-      response.should be_success
-      assigns[:ticket_kind].should == @record
+      expect(response).to be_success
+      expect(assigns[:ticket_kind]).to == @record
     end
   end
 
   describe "update" do
-    it "should succeed with valid attributes" do
+    it "expects to succeed with valid attributes" do
       create_record
       put :update, :id => @record.id, :ticket_kind => @record.attributes
-      response.should redirect_to(ticket_kind_path(@record))
-      assigns[:ticket_kind].should == @record
+      expect(response).to redirect_to(ticket_kind_path(@record))
+      expect(assigns[:ticket_kind]).to == @record
     end
 
-    it "should fail with invalid attributes" do
+    it "expects to fail with invalid attributes" do
       create_record
       put :update, :id => @record.id, :ticket_kind => { :title => nil }
-      response.should be_success
-      assigns[:ticket_kind].should == @record
-      assigns[:ticket_kind].errors.full_messages.first.should =~ /title can't be blank/i
+      expect(response).to be_success
+      expect(assigns[:ticket_kind]).to == @record
+      expect(assigns[:ticket_kind].errors.full_messages.first).to =~ /title can't be blank/i
     end
   end
 
   describe "destroy" do
-    it "should succeed" do
+    it "expects to succeed" do
       create_record
       delete :destroy, :id => @record.id
-      response.should redirect_to(ticket_kinds_path)
-      TicketKind.exists?(@record.id).should be_false
+      expect(response).to redirect_to(ticket_kinds_path)
+      expect(TicketKind.exists?(@record.id)).to be_false
     end
   end
 
