@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe TicketsController do
 
@@ -6,23 +6,23 @@ describe TicketsController do
     login_as_admin
   end
 
-  def mock_ticket(stubs={})
-    @mock_ticket ||= mock_model(Ticket, stubs)
+  def mock_ticket
+    @mock_ticket ||= Ticket.new
   end
 
   describe "GET index" do
     it "assigns all tickets as @tickets" do
-      Ticket.stub(:find).with(:all).and_return([mock_ticket])
+      allow(Ticket).to receive(:ordered).and_return([mock_ticket])
       get :index
-      assigns[:tickets].should == [mock_ticket]
+      expect(assigns[:tickets]).to eq([mock_ticket])
     end
   end
 
   describe "GET show" do
     it "assigns the requested ticket as @ticket" do
-      Ticket.stub(:find).with("37").and_return(mock_ticket)
+      allow(Ticket).to receive(:find).with("37").and_return(mock_ticket)
       get :show, :id => "37"
-      assigns[:ticket].should equal(mock_ticket)
+      expect(assigns[:ticket]).to equal(mock_ticket)
     end
   end
 
