@@ -1,11 +1,11 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe TicketsHelper do
 
   #Delete this example and add some real ones or delete this file
   it "is included in the helper object" do
     included_modules = (class << helper; self; end).send :included_modules
-    included_modules.should include(TicketsHelper)
+    expect(included_modules).to include(TicketsHelper)
   end
 
 end
