@@ -1,11 +1,11 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe UserSessionsHelper do
 
   #Delete this example and add some real ones or delete this file
-  it "should be included in the object returned by #helper" do
+  it "expects to be included in the object returned by #helper" do
     included_modules = (class << helper; self; end).send :included_modules
-    included_modules.should include(UserSessionsHelper)
+    expect(included_modules).to include(UserSessionsHelper)
   end
 
 end
