@@ -1,35 +1,21 @@
-# == Schema Information
-# Schema version: 20100502225937
-#
-# Table name: ticket_kinds
-#
-#  id         :integer         not null, primary key
-#  title      :string(255)
-#  prefix     :string(255)
-#  template   :text
-#  created_at :datetime
-#  updated_at :datetime
-#  subject    :string(255)
-#
-
-require 'spec_helper'
+require 'rails_helper'
 
 describe TicketKind do
   before(:each) do
-    @valid_attributes = Factory.attributes_for :ticket_kind
+    @valid_attributes = attributes_for :ticket_kind
   end
 
-  it "should create a new instance given valid attributes" do
+  it "expects to create a new instance given valid attributes" do
     TicketKind.create!(@valid_attributes)
   end
 
-  it "should set the prefix correctly" do
-    Factory(:ticket_kind, :title => 'Volunteer').prefix.should == 'volunteer'
+  it "expects to set the prefix correctly" do
+    expect(create(:ticket_kind, :title => 'Volunteer').prefix).to == 'volunteer'
   end
 
-  it "should not set prefix if no title is set" do
-    kind = Factory.build(:ticket_kind, :title => nil, :prefix => nil)
-    kind.should_not_receive(:prefix=)
+  it "expects to not set prefix if no title is set" do
+    kind = build(:ticket_kind, :title => nil, :prefix => nil)
+    expect(kind).to_not receive(:prefix=)
     kind.set_prefix
   end
 end
