@@ -1,18 +1,18 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe "/batches/new.html.erb" do
   include BatchesHelper
 
   before(:each) do
-    assigns[:batch] = Factory.build(:batch, :ticket_kind => nil)
-    assigns[:access_ticket_kinds] = [Factory.stub(:ticket_kind)]
-    assigns[:discount_ticket_kinds] = [Factory.stub(:ticket_kind)]
+    assign(:batch, build(:batch, :ticket_kind => nil))
+    assign(:access_ticket_kinds, [build_stubbed(:ticket_kind)])
+    assign(:discount_ticket_kinds, [build_stubbed(:ticket_kind)])
   end
 
   it "renders new batch form" do
     render
 
-    response.should have_tag("form[action=?][method=post]", batches_path) do
+    expect(response).to have_tag("form[action=?][method=post]", batches_path) do
     end
   end
 end
