@@ -1,25 +1,28 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe "/ticket_kinds/new.html.erb" do
   include TicketKindsHelper
 
   before(:each) do
-    assigns[:ticket_kind] = stub_model(TicketKind,
-      :new_record? => true,
-      :title => "value for title",
-      :prefix => "value for prefix",
-      :template => "value for template"
-    )
-    assigns[:eventbrite_tickets] = @eventbrite_tickets = []
+    assign(:ticket_kind, TicketKind.new(
+      title: "value for title",
+      prefix: "value for prefix",
+      subject: "value for subject",
+      template: "value for template"
+    ))
+    assign(:eventbrite_tickets, @eventbrite_tickets = [])
   end
 
   it "renders new ticket_kind form" do
     render
 
-    response.should have_tag("form[action=?][method=post]", ticket_kinds_path) do
-      with_tag("input#ticket_kind_title[name=?]", "ticket_kind[title]")
-      with_tag("input#ticket_kind_prefix[name=?]", "ticket_kind[prefix]")
-      with_tag("textarea#ticket_kind_template[name=?]", "ticket_kind[template]")
-    end
+    expect(view).to render_template(partial: "_form")
+    expect(rendered).to have_selector("form[action=?][method=post]", ticket_kinds_path)
+# TODO
+#    do |node|
+#      expect(node).to have_selector("input#ticket_kind_title[name=?]", "ticket_kind[title]")
+#      expect(node).to have_selector("input#ticket_kind_prefix[name=?]", "ticket_kind[prefix]")
+#      expect(node).to have_selector("textarea#ticket_kind_template[name=?]", "ticket_kind[template]")
+#    end
   end
 end
