@@ -1,14 +1,16 @@
-require 'spec_helper'
+require 'rails_helper'
+require 'support/factory_girl'
+require 'support/disable_register_code'
 
 describe "/tickets/show.html.erb" do
   include TicketsHelper
   before(:each) do
-    assigns[:ticket] = @ticket = Factory(:ticket)
+    assign(:ticket, @ticket = create(:ticket))
   end
 
   it "renders attributes in <p>" do
     render
-    response.should have_text(/#{@ticket.email}/)
-    response.should have_text(/#{@ticket.report}/)
+    expect(response).to have_text(/#{@ticket.email}/)
+    expect(response).to have_text(/#{@ticket.report}/)
   end
 end
