@@ -1,18 +1,5 @@
-# Be sure to restart your server when you modify this file
+# Load the Rails application.
+require File.expand_path('../application', __FILE__)
 
-# Bootstrap the Rails environment, frameworks, and default configuration
-require File.join(File.dirname(__FILE__), 'boot')
-
-Rails::Initializer.run do |config|
-  # See `Gemfile` for list of gem libraries used by this application.
-
-  # Libraries loaded from Ruby itself
-  require 'digest/md5'
-
-  # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-  # Run "rake -D time" for a list of tasks for finding time zone names.
-  config.time_zone = 'UTC'
-
-  # Load custom libraries before "config/initializers" run.
-  $LOAD_PATH.unshift("#{RAILS_ROOT}/lib")
-end
+# Initialize the Rails application.
+Rails.application.initialize!
