@@ -1,6 +1,6 @@
-# Stub SECRETS.eventbrite_data with valid data.
+# Stub Rails.application.secrets.eventbrite_data with valid data.
 def stub_eventbrite_secrets
-  SECRETS.stub!(:eventbrite_data => {
+  Rails.application.secrets.stub!(:eventbrite_data => {
     'app_key'  => '123',
     'user_key' => '456',
     'event_id' => '789',
@@ -8,9 +8,9 @@ def stub_eventbrite_secrets
   })
 end
 
-# Stub SECRETS.eventbrite_data with invalid data.
+# Stub Rails.application.secrets.eventbrite_data with invalid data.
 def stub_invalid_eventbrite_secrets
-  SECRETS.stub!(:eventbrite_data => {
+  Rails.application.secrets.stub!(:eventbrite_data => {
     'app_key'  => 'test',
     'user_key' => 'test',
     'event_id' => 'test',
