@@ -1,13 +1,13 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe "/batches/index.html.erb" do
   include BatchesHelper
 
   before(:each) do
-    assigns[:batches] = [
-      Factory.stub(:batch),
-      Factory.stub(:batch),
-    ]
+    assign(:batches, [
+      build_stubbed(:batch),
+      build_stubbed(:batch),
+    ])
   end
 
   it "renders a list of batches" do
